Homework updates

+++ homework/cs445/hw2_variables.c
@@ -1,5 +1,7 @@
 //hw1 
+// can be used with big.txt, file.txt, and textfile
 #include <stdio.h>
+#include <stdlib.h> // <-- adds exit for part 2
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <sys/wait.h>
@@ -19,14 +21,15 @@ int main(int argc, char const *argv[])
 
 	while(1) {
 
-		signal(SIGUSR1, handler);
-		signal(SIGUSR2, handler);
+		signal(SIGUSR1, handler); //PART 1
+		signal(SIGUSR2, handler); //PART 1
+		signal(SIGALRM, handler);
 
 		if(pid == 0)
 		{
 			fp = fopen(fname, "r");
 			if(fp == NULL){
-				printf("Child cannot find [%s].", fname);
+				printf("\tChild cannot find [%s].", fname);
 				fclose(fp);
 			}
 			else {
@@ -34,34 +37,48 @@ int main(int argc, char const *argv[])
 				fclose(fp);
 				n = st.st_size;
 				pid = getpid();
-				if(n%2) kill(pid, SIGUSR1);
-					else kill(pid, SIGUSR2);
+        		
+				if(n>255) { kill(pid, SIGALRM); exit(0); }
+				else if(n%2) kill(pid, SIGUSR1); //PART 1
+					else kill(pid, SIGUSR2);  //PART 1
+				exit(n);
 			}
-
-				kill(getpid(), 9); //send kill
-				wait(&status);	//double tap
+			kill(getpid(), 9); //send kill if I become a zombie
 		}
 		else
 		{	
-			noRace();
 			printf("Enter filenames until you're happy. 'die' to end.\n");
 			scanf("%s", fname);
 			if(!strcmp(fname,"die")) break;
 			pid = fork();	
-		}
 
+			wait(&status);
+			n = WEXITSTATUS(status);
+			if ( n != 0 && pid != 0) //no file too big && no parent
+    		    printf("\t%s is %d\n", fname, n);
+		}
 	}
 	printf("Bye bye\n\n");
 	return 0;
 }
 
 static void handler(int signo) {
-	if(signo-10) printf("%s is even.\n", fname); else printf("%s is odd.\n", fname);
+	if(signo == SIGALRM)
+		printf("\tFile is too big!!!\n");
+	else
+		printf("\t%s is %s\n",fname, (signo%2)?"odd":"even");
+	return;
 }
 
-void noRace() {
-	int i;
-//	for( i=0;i<3;i++ )
-//		for( i=0;i<100000000;i++ );
-sleep(1);
+void noRace() { // could be used to avoid race condition in part 1 
+  sleep(1);
 }
+/*
+Part 1
+
+Write a program at reads the names of files by reading the standard input. Each time it reads a filename, it creates a new process (using fork) that determines the file's length, using the "stat" system call. If the file length is odd, the child process uses the "kill" system call to send the SIGUSR1 signal to the parent. It the file length is even, the child process sends the SIGUSR2 signal to the parent. When the parent receives the signal, it prints the name of the file and whether its length was odd or even. After printing, the process reads the next line of input from standard input and continues.
+
+PART 2
+
+Use the status of the child (passed back to the parent using the exit library function, see "man 3 exit", and received by the parent using "wait") to report the length of the file back to the parent. Also, print whether odd or even as reported thtough the signal handler above. In addition, if the file length exceeds 255 bytes, use the SIGALRM signal to notify the parent process, so that it can print an error message, instead of odd/even. When the file size exceeds 255 bytes, an length of 0 can be returned, instead of the file length.
+*/
